:construction: Add an exit button and aply state variables

+++ src/GameState.java
@@ -5,7 +5,7 @@ public abstract class GameState {
 	protected int mixedFollowRate;
 	protected int speed;
 	protected int cherryChance;
-	protected float boostedTime;
+	protected int boostedTime;
 	protected int lives;
 	protected GameObject[][] map;
 	protected float pointMultiplier;
@@ -17,7 +17,7 @@ public abstract class GameState {
 	int getMixedFollowRate() {return this.mixedFollowRate;}
 	int getSpeed() {return this.speed;}
 	int getCherryChance() {return this.cherryChance;}
-	float getBoostTime(){return this.boostedTime;}
+	int getBoostTime(){return this.boostedTime;}
 	int getLives(){return this.lives;}
 	GameObject[][] getMap(){return this.map;}
 	float getPointMultiplier() {return this.pointMultiplier;}
